Add select mode that does nothing when clicked
The select mode have no purpose in a PC, but it's useful in touch panel
UI devices since the user have no means to just select and see
information of a random cell without it.

It's also a safe mode to prevent accidentally wasting working power to
cultivating or something.