Add button tip window to show functionality of the button
The tip window shows only when the mouse cursor is hovering over
the button, which makes touch panels unable to show them at all.
This issue may be hard to resolve if we keep the UI shared between
the mouse and the touch panel.

Also we made the variable clickMode a function object that can
explain about itself, and renamed it to clickCallback.
This kind of dynamic property modification is very handy in
JavaScript.