Make update rate of the game mostly constant regardless of CPU power
The drawing method is costly process and the time to redraw may vary
depending on devices.  Nevertheless, we want to keep the simulation
rate mostly constant, regardless of running environment.  So we update
the game with real time interval, while repainting is only performed as
often as the device can.

If the device's CPU could not process game update routine on time,
the game wouldn't be playable, since the game logic assumes the CPU
can handle it in real time.  Probably we should check to see this is
the case.