Make potatoes degrade by continuous cropping
You can crop corns in between potatoes to prevent diseases induced
by continuous cropping.  I think it's called rotation in farming
terminology, but I'm not confident.

We had to introduce quality to crops and potatoPest to cells in
order to realize degradation.  The simulated rate of degradation is
suboptimal.  We could insert another delay element to the modulation
of quality for more realistic result.