Fix scale of Crop.amount to match object oriented model
Previously, we have been using 0 between 1 to denote that no crop is
growing, but now that we have moved the definition of amount to Crop
class member, we don't need the range for special meanings.
The more natural and intuitive definition of amount is to start off
at 0 and harvest at 1.